# BE/openapi.snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BE",
    "version": "1.0.0"
  },
  "paths": {
...
    "/observability/sessions/{cli_invocation_id}/logs": {
      "get": {
        "summary": "Tail Session Logs",
        "description": "Return a page of JSONL log lines for a CLI invocation.",
        "operationId": "tail_session_logs_observability_sessions__cli_invocation_id__logs_get",
        "parameters": [
          {
            "name": "cli_invocation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Cli Invocation Id"
            }
          },
          {
            "name": "tail",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "1..2000, default 200",
              "default": 200,
              "title": "Tail"
            },
            "description": "1..2000, default 200"
          },
          {
            "name": "after_offset",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "resume byte offset; overrides tail",
              "title": "After Offset"
            },
            "description": "resume byte offset; overrides tail"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}